Add registration and login

+++ utils/deleteUploadedFile.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const path = require('path');
 
 const deleteUploadedFile = (filename) => {
-	fs.unlinkSync((path.join(process.cwd() + '/public/upload/' + filename)));
+	if(filename) fs.unlinkSync((path.join(process.cwd() + '/public/upload/' + filename)));
 };
 
 module.exports = deleteUploadedFile;
